babiesApi/requirements.txt
asgiref==3.2.7
astroid==2.3.3
Django==3.0.5
django-cors-headers==3.2.1
django-guardian==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
